pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "base"
description = "Base Project"
authors = [
    {name = "me", email = "mail@example.me"},
]
dependencies = [
    "compose>=1.4.8",
    "scikit-learn>=1.3.0",
    "torchvision>=0.15.2",
    "tqdm>=4.65.0",
    "wandb>=0.15.5",
    "rtpt>=0.0.4",
    "PyWavelets>=1.4.1",
    "dtcwt>=0.12.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
	"Programming Language :: Python :: 3",
	"License :: OSI Approved :: MIT License",
	"Operating System :: OS Independent",
]
dynamic = ["version"]

[project.optional-dependencies]
cuda = [
]
[tool.pdm]
version = {use_scm = true}

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.3.1",
]
dev = [
    "black>=23.3.0",
]

[[tool.pdm.source]]
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
name = "jax_cuda_releases"
type = "find_links"

[[tool.pdm.source]]
url = "https://download.pytorch.org/whl/cpu"
name = "pytorch"
type = "find_links"

[tool.pdm.scripts]
pre_example = "docker compose up --no-recreate -d"
example.cmd = "docker compose exec -e CUDA_VISIBLE_DEVICES -e JAX_PLATFORMS -w /base/examples main python"
test.cmd = "docker compose exec -e CUDA_VISIBLE_DEVICES='' -e JAX_PLATFORMS=cpu -w /base main python -m pytest"
exec.cmd = "docker compose exec -e CUDA_VISIBLE_DEVICES -e JAX_PLATFORMS main python"
black.cmd = "docker compose exec -w /senn main python -m black ."

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
